Allow Dispatcher to create a clean request by itself:
As part of untangling the Core::Context object, we first allow the
Dispatcher class to create its own request inside the dispatching
process.

It does this for each app, in the loop, because the request is
dependent on the app. Creating the app is measured to be a very
cheap operation.

The other way would be to make the "app" attribute in the request
class a read-write, and then set it, which we prefer not to do. Or,
alternatively, we could keep the app alive all the time and pass it
as a parameter between all functions. Eww.