Add preliminary session attribute to Core::App:
This is a tricky way. Stay with me.

We want to move the session from Core::Context to Core::App. In
order to do this gradually, we started by creating a session attribute
in Core::App and update it with the session object when we create it
in the DSL for the context.

This means Context and App now share the same session object across
two attributes. So far so good.

Next, we want Engines (which are related to an App) to be able to
reach the session. Core::Role::Template actually uses this. So, we
added a session attribute to the role for engines (Core::Role::Engine).

The session setup (called from the DSL, right after it creates a
session using the Context object) will copy the session object to
all engines.

We can now start removing more Core::Context appearances:
* We remove "setup_context" call from the Dispatcher.
* We remove "setup_context" method in Core::App.
* We remove the trigger for it from "context" attribute in App.
* We remove the "context" attribute from Core::Role::Engine.
* We remove the context triggers test.

But now Core::Role::Template has a problem. It has been using the
context in order to reach the updated configuration in the app
(context->app->config) in order to populate $tokens->{'settings'}.

In order to handle this, we add an attribute, "settings", which we
then populate ourselves with the config right before we process the
template. That way it always has the updated config in the template
object, and $tokens->{'settings'} works without the context.

There is still one problem which we're not sure when it appeared,
and might have always existed. Honestly, we're just too tired to check.

The handlers (AutoPage/File) have an attribute "app", which stores a
reference to the app. The app stores the config, which includes the
"route_handlers" key, which then includes all objects of the handlers.

This means that App->Config->Route Handlers->App. Full circle,
circular referencing bonanza.

We will need to fix it in a later commit somehow.