Provide Serializer::Mutable with more data, fixing last failing test:
The last failing test was erroring because the mutable serializer
(Serializer::Mutable) was trying to fetch header information in order
to dynamically figure out the content type.

Those headers are available in the request, which isn't exposed
to serializers. It did it using the context, which is no longer
avaliable.

This commit adds only those specific headers to the serializer
(without creating an attribute for them, just keys in the object
hash) so the serializer could figure this out.

It's in interim solution which I hope to remove later on, when we
go over the mutable serializer in general. It also has hard-coded
support for specific serializers and deserializers, which I want
to also change, along with it.